+++ src/Exceptions/UnreadableResponseException.php
@@ -2,8 +2,6 @@
 
 namespace Zoho\Crm\Exceptions;
 
-use Exception;
-
 class UnreadableResponseException extends Exception
 {
     /** @var string The exception message */