+++ CHANGELOG.md
@@ -14,6 +14,7 @@ https://github.com/tristanjahier/zoho-crm-php/compare/0.5.0...master
 - Dropped support for PHP 7.
 - Dropped support for Guzzle 6.
 - Dropped support for Doctrine Inflector 1.
+- `Zoho\Crm\V2\ResponseParser` will now throw an `UnreadableResponseException` when the API response body cannot be parsed.
 
 ### Removed